FU-79 feat: productService 인터페이스 삭제
